Fix program from freezing up
server.go:
The program could end up freezing up due to continuously trying
to connect to a peer in the loop not allowing RPC-requests to
happen.

cpuminer.go:
Disabled the code that would prevent mining without a peer. Since
the network is small a single node should be allowed to mine blocks
on its own.

These are both put as temporary solutions.