Update block serialization and database format
Transaction abstraction is now reflected on the new serialization
format. Each block now looks like this both over the wire and on
disk.

.--------------.
| Block Header | <--- Standard block header
|--------------|
| Tx Count (N) | <--- Number of transactions in the block
|--------------|
| Tx Header #0 | <--- A transaction header contains the type of
|--------------|      transaction contained and where in the
| Tx Header #1 |      block it's located and the size.
|--------------|
|      .       |
|      .       |
|--------------|
| Tx Header #N |
|--------------|
|  Tx Body #1  | <--- The bytes containing the transaction as
|--------------|      pointed to by the coresponding transaction
|  Tx Body #2  |      header.
|--------------|
|      .       |
|      .       |
|--------------|
|  Tx Body #N  |
'--------------'

MsgTx have gotten a new method called MsgDecodeFixed which acts
just like MsgDecode expect that it doesn't consume the entire
input. It takes a new size paramter which restricts the number of
bytes that will become a part of the Data field of MsgTx. This is
needed since the size of the transaction cannot be deduced from
reading the bytes and have to be passed in. It's always assumed
that there will be at least 8 bytes to accomedate the MsgTx header.
This may change in the future as this header will probably merge
with the already exsiting transaction header when serialized.

msgblock_test.go has also been updated with new test cases to work
with this new format.

Bug fixes:

- In mining.go there was a bug where the "coinbase" transaction
  were about twice the size it should be when serialized. This
  was fixed by having the buffer have an initial length of 0.

- In msgtx.go there was a bug in the Copy method were no bytes
  were being copied and in the wrong direction.